+++ src/lib.rs
@@ -8,29 +8,27 @@
 //!  stores metadata like its numeric code, full name and symbol.
 //!
 //! ```edition2018
-//! use rusty_money::{money, Money, Currency};
+//! use rusty_money::{Money, Currency};
 //! 
 //! let currency = Currency::new("USD", 2, 2);
 //! let one_dollar = Money::from_minor(100, &currency); // 1 USD
 //! 
-//! // You can also initialize formattable currency, which can be printed in a few ways.
-//!
 //!   
 //! // Money can be initialized in a few ways:
 //! use rusty_money::Locale::*;
 //! use rusty_money::{IsoCurrency};
 //! use rusty_money::Iso::*;
 //!
-//! money!(2000, "USD");                               // 2000 USD
-//! money!("2000.00", "USD");                          // 2000 USD
-//! Money::from_major(2000, IsoCurrency::get(USD));    // 2000 USD
-//! Money::from_minor(200000, IsoCurrency::get(USD));  // 2000 USD
-//! Money::from_str("2,000.00", "USD").unwrap();       // 2000 USD
+//! Money::from_stringable(2000, "USD").unwrap();       // 2000 USD
+//! Money::from_stringable("2000.00", "USD").unwrap();  // 2000 USD
+//! Money::from_major(2000, IsoCurrency::get(USD));     // 2000 USD
+//! Money::from_minor(200000, IsoCurrency::get(USD));   // 2000 USD
+//! Money::from_stringable("2,000.00", "USD").unwrap(); // 2000 USD
 //!
 //!
 //! // Money objects with the same Currency can be compared:
-//! let hundred = money!(100, "USD");
-//! let thousand = money!(1000, "USD");
+//! let hundred = Money::from_stringable(100, "USD").unwrap();
+//! let thousand = Money::from_stringable(1000, "USD").unwrap();
 //! println!("{}", thousand > hundred);     // false
 //! println!("{}", thousand.is_positive()); // true
 //! ```
@@ -47,19 +45,19 @@
 //! * [Half Even](https://en.wikipedia.org/wiki/Rounding#Round_half_even) (default)
 //!
 //! ```edition2018
-//! use rusty_money::{money, Money, IsoCurrency, Round};
+//! use rusty_money::{Money, IsoCurrency, Round};
 //!
 //! // Money can be added, subtracted, multiplied and divided:
-//! money!(100, "USD") + money!(100, "USD");        // 200 USD
-//! money!(100, "USD") - money!(100, "USD");        // 0 USD
-//! money!(1, "USD") * 3;                           // 3 USD
-//! money!(3, "USD") / 3;                           // 0.333333333... USD
+//! Money::from_stringable(100, "USD").unwrap() + Money::from_stringable(100, "USD").unwrap(); // 200 USD
+//! Money::from_stringable(100, "USD").unwrap() - Money::from_stringable(100, "USD").unwrap(); // 0 USD
+//! Money::from_stringable(1, "USD").unwrap() * 3;                                             // 3 USD
+//! Money::from_stringable(3, "USD").unwrap() / 3;                                             // 0.333333333... USD
 //!
 //! // Money can be rounded by calling the round function:
-//! let usd = money!("-2000.005", "USD");           // 2000.005 USD
-//! usd.round(2, Round::HalfEven);                  // 2000.00 USD
-//! usd.round(2, Round::HalfUp);                    // 2000.01 USD
-//! usd.round(0, Round::HalfUp);                    // 2000 USD
+//! let usd = Money::from_stringable("-2000.005", "USD").unwrap(); // 2000.005 USD
+//! usd.round(2, Round::HalfEven);                                 // 2000.00 USD
+//! usd.round(2, Round::HalfUp);                                   // 2000.01 USD
+//! usd.round(0, Round::HalfUp);                                   // 2000 USD
 //!```
 //!
 //! ## Formatting
@@ -69,11 +67,11 @@
 //! accepts a more detailed set of parameters.
 //!
 //! ```edition2018
-//! use rusty_money::{money, Money, IsoCurrency};
+//! use rusty_money::{Money, IsoCurrency};
 //!
 //! // Money objects can be pretty printed, with appropriate rounding and formatting:
-//! let usd = money!("-2000.009", "USD");
-//! let eur = money!("-2000.009", "EUR");
+//! let usd = Money::from_stringable("-2000.009", "USD").unwrap();
+//! let eur = Money::from_stringable("-2000.009", "EUR").unwrap();
 //! println!("{}", usd); // -$2,000.01
 //! println!("{}", eur); // -€2.000,01;
 //! ```
@@ -84,13 +82,13 @@
 //! to another.
 //!
 //! ```edition2018
-//! use rusty_money::{money, Money, IsoCurrency, Exchange, ExchangeRate};
+//! use rusty_money::{Money, IsoCurrency, Exchange, ExchangeRate};
 //! use rusty_money::Iso::*;
 //! use rust_decimal_macros::*;
 //!
 //! // Convert 1000 USD to EUR at a 2:1 exchange rate.
 //! let rate = ExchangeRate::new(IsoCurrency::get(USD), IsoCurrency::get(EUR), dec!(0.5)).unwrap();
-//! rate.convert(money!(1000, "USD")); // 500 EUR
+//! rate.convert(Money::from_stringable(1000, "USD").unwrap());                                     // 500 EUR
 //!
 //! // An Exchange can be used to store ExchangeRates for later use
 //! let mut exchange = Exchange::new();